refactor(A2-2064): stop creating duplicate get representation formatter functions
